move mocks to 'syncer/pkg/mock/'
